+++ controllers/products.js
@@ -6,10 +6,17 @@ exports.getAll = async (req, res) => {
         const videoId = req.params.videoId;
         const products = await Product.find({ video_id: videoId });
 
-        return res.status(200).json({'products': products});
+        return res.status(200).json({
+            'status': 'success',
+            'message': 'products successfully fetched',
+            'data': products
+        });
     }
     catch (err) {
-        return res.status(500).json({ 'message': err.message });
+        return res.status(500).json({
+            'status': 'failed',
+            'message': err.message
+        });
     }
 }
 
@@ -18,6 +25,7 @@ exports.create = async (req, res) => {
         const videoId = req.params.videoId;
         const id = `product-${uuidv4()}`;
         const { title, price, url_image, url_product } = req.body;
+
         const product = new Product({
             _id: id,
             video_id: videoId,
@@ -26,21 +34,44 @@ exports.create = async (req, res) => {
             url_image,
             url_product
         });
+
         const newProduct = await product.save();
 
-        return res.status(201).json({'product': newProduct});
+        return res.status(201).json({
+            'status': 'success',
+            'message': 'product successfully created',
+            'data': newProduct
+        });
     }
     catch (err) {
-        return res.status(500).json({ 'message': err.message });
+        return res.status(500).json({
+            'status': 'failed',
+            'message': err.message
+        });
     }
 }
 
 exports.onDelete = async (req, res) => {
     try {
+        const product = await Product.findById(req.params.productId);
+
+        if (!product) {
+            return res.status(404).json({
+                'status': 'failed',
+                'message': 'product not found'
+            });
+        }
+
         await Product.findByIdAndDelete(req.params.productId);
 
-        return res.status(200).json({'message': 'Product deleted successfully'});
-    } catch (error) {
-        return res.status(500).json({ 'message': error.message });
+        return res.status(200).json({
+            'status': 'success',
+            'message': 'product successfully deleted'
+        });
+    } catch (err) {
+        return res.status(500).json({
+            'status': 'failed',
+            'message': err.message
+        });
     }
 }
